chore: move `makeInputJsFieldsReadable` to `input_schema` package
